Moving classes to main package and including icon.
