+++ src/main/java/org/netbeans/modules/liberty/main/LibertyInstanceProvider.java
@@ -1,4 +1,4 @@
-package org.netbeans.modules.liberty;
+package org.netbeans.modules.liberty.main;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,11 +9,13 @@ import org.netbeans.spi.server.ServerInstanceProvider;
 
 public class LibertyInstanceProvider implements ServerInstanceProvider {
 
+    private static final String serverInstanceName = "WebSphere Liberty";
+    
     @Override
     public List<ServerInstance> getInstances() {
         List<ServerInstance> instances = new ArrayList<ServerInstance>();
         ServerInstance instance = ServerInstanceFactory.createServerInstance(
-                new LibertyInstanceImplementation(this, "Liberty", "Liberty Instance", true));
+                new LibertyInstanceImplementation(this, serverInstanceName, serverInstanceName, true));
         instances.add(instance);
         return instances;
     }
